Move logic for Module names into core
