+++ packages/hekla-cli/src/commands/analyze.js
@@ -1,4 +1,7 @@
+const fs = require('fs');
 const path = require('path');
+
+const Analyzer = require('hekla-core').Analyzer;
 
 module.exports = function analyze(cmd) {
   const { single } = cmd;
@@ -11,6 +14,11 @@ module.exports = function analyze(cmd) {
     const filePath = path.resolve(process.cwd(), single);
 
     console.log('Resolved:', filePath);
+
+    const analyzer = new Analyzer();
+    analyzer.setInputFileSystem(fs);
+    const module = analyzer.createModule('', filePath);
+    analyzer.processModule(module);
   }
 
   // let config;
